Rename `kc_reader` to `kc_access`
