+++ kpi/deployment_backends/kobocat_backend.py
@@ -16,7 +16,7 @@ from rest_framework import exceptions, status
 from rest_framework.authtoken.models import Token
 
 from base_backend import BaseDeploymentBackend
-from .kc_reader.utils import instance_count
+from .kc_access.utils import instance_count
 
 
 class KobocatDeploymentException(exceptions.APIException):
@@ -394,7 +394,7 @@ class KobocatDeploymentBackend(BaseDeploymentBackend):
     def _submission_count(self):
         _deployment_data = self.asset._deployment_data
         id_string = _deployment_data['backend_response']['id_string']
-        # avoid migrations from being created for kc_reader mocked models
+        # avoid migrations from being created for kc_access mocked models
         # there should be a better way to do this, right?
         return instance_count(xform_id_string=id_string,
                               user_id=self.asset.owner.pk,
