+++ kpi/management/commands/copy_kc_profile.py
@@ -3,7 +3,7 @@ from django.core.management.base import BaseCommand, CommandError
 from django.contrib.auth.models import User
 
 from hub.models import ExtraUserDetail
-from kpi.deployment_backends.kc_reader.utils import get_kc_profile_data
+from kpi.deployment_backends.kc_access.utils import get_kc_profile_data
 
 class Command(BaseCommand):
     option_list = BaseCommand.option_list + (
